Admin staff page

+++ src/main/java/kea/kinoBackend/project/controller/ReservationController.java
@@ -3,8 +3,14 @@ package kea.kinoBackend.project.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import kea.kinoBackend.project.dto.ReservationDTO;
 import kea.kinoBackend.project.service.ReservationService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,9 +19,11 @@ import java.util.List;
 @RequestMapping("/reservations")
 public class ReservationController {
     private ReservationService reservationService;
+    private JwtDecoder jwtDecoder;
 
-    public ReservationController(ReservationService reservationService) {
+    public ReservationController(ReservationService reservationService, JwtDecoder jwtDecoder) {
         this.reservationService = reservationService;
+        this.jwtDecoder = jwtDecoder;
     }
 
     @Operation(summary = "Get all reservations", description = "Get a list of all reservations")
@@ -34,6 +42,29 @@ public class ReservationController {
 
     @Operation(summary = "Add a new reservation", description = "Add a new reservation")
     @PreAuthorize("hasAnyAuthority('ADMIN','STAFF','USER')")
+    @GetMapping("/user")
+    public List<ReservationDTO> getReservationsByUser(@RequestHeader("Authorization") String bearerToken) {
+        // Remove the 'Bearer ' prefix from the token
+        String token = bearerToken.substring(7);
+
+        // Decode the token
+        Jwt jwt = jwtDecoder.decode(token);
+        System.out.println(jwt);
+
+        // Extract the username from the token.
+        String currentUsername = jwt.getSubject();;
+
+        if (currentUsername == null) {
+            throw new IllegalArgumentException("Username claim could not be found in the Jwt.");
+        }
+
+        System.out.println("Current user: " + currentUsername);
+
+        // Fetch reservations for the currently authenticated user
+        return reservationService.getReservationsByUser(currentUsername);
+    }
+
+    @PreAuthorize("hasAnyAuthority('ADMIN','STAFF','USER')")
     @PostMapping
     public ReservationDTO addReservation(@RequestBody ReservationDTO request) {
         return reservationService.addReservation(request);
@@ -42,7 +73,35 @@ public class ReservationController {
     @Operation(summary = "delete a reservation", description = "delete a reservation")
     @PreAuthorize("hasAnyAuthority('ADMIN','STAFF','USER')")
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteReservation(@PathVariable int id) {
-        return reservationService.deleteReservation(id);
+    public ResponseEntity deleteReservation(@PathVariable int id, @RequestHeader("Authorization") String bearerToken) {
+        //Allow admin and staff to delete without checking token
+        if (SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("ADMIN") || a.getAuthority().equals("STAFF"))) {
+            return reservationService.deleteReservation(id);
+        }
+
+        ReservationDTO reservation = reservationService.getReservationById(id);
+
+        // Remove the 'Bearer ' prefix from the token
+        String token = bearerToken.substring(7);
+
+        // Decode the token
+        Jwt jwt = jwtDecoder.decode(token);
+        System.out.println(jwt);
+
+        // Extract the username from the token.
+        String currentUsername = jwt.getSubject();;
+
+        if (currentUsername == null) {
+            throw new IllegalArgumentException("Username claim could not be found in the Jwt.");
+        }
+
+        System.out.println("Current user: " + currentUsername);
+
+        // Check if the currently authenticated user is the same as the user who made the reservation
+        if (reservation.getUsername().equals(currentUsername)) {
+            return reservationService.deleteReservation(id);
+        } else {
+            return new ResponseEntity<>("You are not authorized to delete this reservation", HttpStatus.UNAUTHORIZED);
+        }
     }
 }
